# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Zenuniverse Mobile API
  version: 1.0.0
  description: 'Kontrak dari sumber lokal routes/api.php, controller, resource, middleware
    dan services.

    Bukan bukti deployment/staging atau login Google native live. Semua respons sukses
    HTTP 200.

    Request snake_case; respons umumnya camelCase. Tidak ada refresh token atau endpoint
    avatar.

    '
servers:
- url: https://zenuniverse.id/api/v1
  description: Live integration API; native Google login Android/iOS still requires
    device verification.
security:
- learnerBearer: []
paths:
...
  /learning/attempts:
    post:
      operationId: submitAttempt
      summary: Verifikasi satu langkah dengan UUID idempotensi; throttle 60/menit.
      description: 'Simpan attempt_id dan JSON payload yang sama untuk retry. UUID
        sama dengan payload

        berbeda menghasilkan attempt_conflict. Replay mengembalikan xpAwarded historis,

        bukan pemberian XP ulang; progress tetap snapshot terbaru. Revisi usang ditolak

        sebelum replay. completed menandakan completion lesson, bukan hanya step.

        Jawaban salah adalah sukses HTTP 200 dengan result.correct=false.

        '
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttemptRequest'
      responses:
        '200':
          description: Hasil verifikasi dan progress authoritative.
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/Envelope'
                - type: object
                  properties:
                    data:
                      allOf:
                      - $ref: '#/components/schemas/CompletionResult'
                      - type: object
                        required:
                        - result
                        properties:
                          result:
                            type: object
                            required:
                            - correct
                            - consumeHeart
                            - feedback
                            properties:
                              correct:
                                type: boolean
                              consumeHeart:
                                type: boolean
                              feedback:
                                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Validation'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/ServerError'
  /learning/lessons/{lesson}/complete:
    post:
      operationId: completeLesson
      summary: Rekonsiliasi completion lesson; tidak melewati langkah belum selesai
        (60/menit).
      description: Semua langkah harus selesai untuk reward. Jika belum, HTTP 200
        dengan completed=false dan xpAwarded=0; tidak ada attempt_id pada endpoint
        ini.
      parameters:
      - $ref: '#/components/parameters/Lesson'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - content_revision
              properties:
                content_revision:
                  type: integer
                  minimum: 1
      responses:
        '200':
          description: Status completion dan snapshot progress.
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/Envelope'
                - type: object
                  properties:
                    data:
                      $ref: '#/components/schemas/CompletionResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Validation'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/ServerError'
components:
  securitySchemes:
    learnerBearer:
      type: http
      scheme: bearer
      description: Personal access token Sanctum dengan ability learner; bukan JWT
        Google atau cookie session.
  parameters:
...
    Lesson:
      name: lesson
      in: path
      required: true
      schema:
        type: string
  responses:
    Unauthorized:
      description: unauthenticated (absen, kedaluwarsa, dicabut) atau invalid_google_token
        pada login.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: forbidden (ability bukan learner) atau lesson_locked.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: not_found; ID hilang atau konten bukan published.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Conflict:
      description: account_link_required (login), content_changed, attempt_conflict
        (attempt), no_hearts (attempt non-concept).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Validation:
      description: validation_failed dengan fields, atau invalid_step untuk step di
        lesson lain (fields kosong).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    RateLimited:
      description: rate_limited; hormati Retry-After bila dikirim.
      headers:
        Retry-After:
          schema:
            type: string
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    ServerError:
      description: server_error; gunakan requestId untuk diagnosis, jangan log token/jawaban
        sensitif.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
    Envelope:
      type: object
      required:
      - data
      - meta
      description: 'JSON resource Laravel dibungkus data; Cache-Control: no-store
        dan X-Request-ID pada respons API.'
      properties:
        data: {}
        meta:
          type: object
          required:
          - requestId
          properties:
            requestId:
              type: string
              format: uuid
...
    ErrorEnvelope:
      type: object
      required:
      - error
      - meta
      properties:
        error:
          type: object
          required:
          - code
          - message
          - fields
          properties:
            code:
              type: string
            message:
              type: string
            fields:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
        meta:
          type: object
          required:
          - requestId
          properties:
            requestId:
              type: string
              format: uuid
...
    AttemptRequest:
      type: object
      required:
      - lesson_id
      - step_id
      - attempt_id
      - content_revision
      - answer
      properties:
        lesson_id:
          type: string
          maxLength: 160
        step_id:
          type: string
          maxLength: 160
        attempt_id:
          type: string
          format: uuid
        content_revision:
          type: integer
          minimum: 1
        answer:
          type: object
          maxProperties: 32
          required:
          - type
          description: Field jenis jawaban bersifat sometimes pada validasi request;
            missing/mismatch dapat menghasilkan correct=false, bukan selalu 422. concept
            tidak memeriksa acknowledged.
          properties:
            type:
              type: string
              enum:
              - concept
              - quiz
              - blockly
              - code-fill
              - code
            acknowledged:
              type: boolean
            optionId:
              type: string
              maxLength: 160
            commands:
              type: array
              maxItems: 500
              items: {}
            answers:
              type: object
              maxProperties: 500
              additionalProperties:
                type: string
                maxLength: 2000
              description: Map blank ID dari content.blanks ke string jawaban.
            code:
              type: string
              maxLength: 20000
    CompletionResult:
      type: object
      required:
      - xpAwarded
      - completed
      - progress
      properties:
        xpAwarded:
          type: integer
        completed:
          type: boolean
        progress:
          $ref: '#/components/schemas/Progress'
    Progress:
      type: object
      required:
      - totalXp
      - updatedAt
      - dailyXp
      - dailyXpDate
      - dailyGoalClaimedDate
      - currentStreak
      - longestStreak
      - lastActivityDate
      - hearts
      - heartsCapacity
      - heartsRegenMinutes
      - nextHeartInMs
      - badges
      - level
      - dailyGoal
      - completedLessonIds
      - completedStepIds
      - courseProgress
      properties:
        totalXp:
          type: integer
        updatedAt:
          type: integer
          format: int64
          description: Unix epoch milliseconds; bukan ISO timestamp.
        dailyXp:
          type: integer
        dailyXpDate:
          type:
          - string
          - 'null'
          format: date
        dailyGoalClaimedDate:
          type:
          - string
          - 'null'
          format: date
        currentStreak:
          type: integer
        longestStreak:
          type: integer
        lastActivityDate:
          type:
          - string
          - 'null'
          format: date
        hearts:
          type: integer
        heartsCapacity:
          type: integer
        heartsRegenMinutes:
          type: integer
        nextHeartInMs:
          type:
          - integer
          - 'null'
          format: int64
        badges:
          type: array
          items:
            type: string
        level:
          type: object
          required:
          - current
          - currentXp
          - xpIntoLevel
          - xpToNextLevel
          - percent
          properties:
            current:
              type: object
              required:
              - level
              - name
              - minXp
              - nextXp
              properties:
                level:
                  type: integer
                name:
                  type: string
                minXp:
                  type: integer
                nextXp:
                  type: integer
            currentXp:
              type: integer
            xpIntoLevel:
              type: integer
            xpToNextLevel:
              type: integer
            percent:
              type: integer
        dailyGoal:
          type: object
          required:
          - progress
          - percent
          - claimed
          - remaining
          properties:
            progress:
              type: integer
            percent:
              type: integer
            claimed:
              type: boolean
            remaining:
              type: integer
        completedLessonIds:
          type: array
          items:
            type: string
        completedStepIds:
          type: array
          items:
            type: string
        courseProgress:
          type: array
          items:
            type: object
            required:
            - courseId
            - title
            - completed
            - total
            - percent
            properties:
              courseId:
                type: string
              title:
                type: string
              completed:
                type: integer
              total:
                type: integer
              percent:
                type: integer